Back_end/requirements.txt
amqp==5.1.0
asgiref==3.5.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.6
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.12
django-celery-beat==2.2.1
django-cors-headers==3.11.0
django-rest-auth==0.9.5
django-timezone-field==4.2.3
django-tinymce==3.4.0
djangorestframework==3.13.1
idna==3.3
kavenegar==1.1.2
kombu==5.2.4
Pillow==9.1.0
prompt-toolkit==3.0.29
psycopg2==2.9.3
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
